refactor(ui): move hooks into one directory in `shared/` dir
- most were in `shared/`, but one was in `components/` (despite not being a component), and another was created in a separate `hooks/` dir
  - put them all in `shared/` for consistency right now
    - might be good to put them into `shared/hooks/`, but for now make it consistent at least
      - also might not be necessary since they do all start with `use-` by convention
  - fix naming of `use-local-storage` which previously did not use any casing, but should use kebab-case for consistency with the rest of the UI codebase
