+++ ui/src/app/shared/use-collect-event.ts
@@ -1,5 +1,5 @@
 import {useEffect} from 'react';
-import {services} from '../../shared/services';
+import {services} from './services';
 
 export function useCollectEvent(name: string) {
     useEffect(() => {
