+++ ui/src/app/workflows/components/workflow-logs-viewer/workflow-logs-viewer.tsx
@@ -12,7 +12,7 @@ import {ErrorNotice} from '../../../shared/components/error-notice';
 import {InfoIcon, WarningIcon} from '../../../shared/components/fa-icons';
 import {Links} from '../../../shared/components/links';
 import {Context} from '../../../shared/context';
-import {useLocalStorage} from '../../../shared/hooks/uselocalstorage';
+import {useLocalStorage} from '../../../shared/use-local-storage';
 import {getPodName, getTemplateNameFromNode} from '../../../shared/pod-name';
 import {ScopedLocalStorage} from '../../../shared/scoped-local-storage';
 import {services} from '../../../shared/services';
